Working power functions and now getting 12bit turns.

+++ Quadruped_Bot_Actuation_Code/include/SPI_NCDR_FCTNS.h
@@ -12,7 +12,7 @@
 #include <SPI.h>
 
 // Function declarations
-uint16_t readEncoderPosition(uint8_t channel);
+float readEncoderPosition(uint8_t channel);
 int16_t readTurns(uint8_t channel);
 void resetEncoder(uint8_t channel);
